Add files via upload
We have two monkeys, a and b, and two boolean values (aSmile and bSmile) indicate if each is smiling. We are in trouble if they are both smiling or if neither of them is smiling. Write a program asking the user if monkeys are smiling and print an alert if we are in trouble (see java.util.Scanner).