Add files via upload
The computer selects a random number between [0, 100] representing the price of an imaginary item. Two users are asked to guess the right price. The user which proposed the number closer to the right price wins. Note that Math.abs(n) returns the absolute value of a number (see java.util.random.RandomGenerator, java.util.Scanner).