Add files via upload
Consider the series of numbers beginning at start and running up to but not including end. For example start=1 and end=5 gives the series 1, 2, 3, 4. Return a new String[] array (see Integer.toString()) containing:

Fizz for multiples of 3;
Buzz for multiples of 5;
FizzBuzz for multiple of both 3 and 5;
the String version of the number in all the other cases.