Add files via upload
Given a String[] and a String return true if the String is contained within the String[]. If the String is not present return false. Write two different versions: (a) uses brute-force search (i.e., which is slow! because tests all elements of the array); (b) sorts the array and uses binary search (which is much faster!) (see java.util.Arrays)

Both versions should not modify the original array!

Examples:

containsBruteForce(["car", "boat", "bike", "plane"], "bike") → true