Add files via upload
Write a class named Circle representing a Circle on a 2D plane. Internally, the class uses a Point object and an int value for representing the center and the radius of the Circle, respectively. The class provides the following methods:

public Circle(Point center, int radius) creating the circle.
getters and setters.
public double getPerimeter() returning the perimeter of the circle.
public double getArea() returning the area of the circle.
public boolean contains(Point point) returning true if point is contained within the circle.
public void translate(int dx, int dy) moving the circle on the 2D plane. dx and dy are the x and y components of the translation vector.
public String toString().